Start using runNodule node metadata in activity log.
